% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_elementary_basis.R
\name{compute_hinge}
\alias{compute_hinge}
\title{Compute the hinge function}
\usage{
compute_hinge(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
Given a numeric vector, this function evaluates the hinge function
\eqn{x -> max(x, 0)} for each component.
}
