% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_elementary_basis.R
\name{compute_indicator}
\alias{compute_indicator}
\title{Compute the indicator function}
\usage{
compute_indicator(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
Given a numeric vector, this function returns the logical vector indicating
whether or not each component is nonnegative.
}
