% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mse.R
\name{compute_mse}
\alias{compute_mse}
\title{Compute the mean squared error of an estimator}
\usage{
compute_mse(y, y_hat)
}
\arguments{
\item{y}{A numeric observation vector.}

\item{y_hat}{A numeric estimate vector of the same length.}
}
\description{
Compute the mean squared error of an estimator
}
\examples{
compute_mse(c(1.0, 2.3, 0.8), c(1.3, 1.9, 0.7))
}
