% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lasso_problem_soln.R
\name{get_lasso_problem_soln}
\alias{get_lasso_problem_soln}
\title{Find the solution to the LASSO problem of an estimation method}
\usage{
get_lasso_problem_soln(
  X_design,
  y,
  s,
  method,
  V = Inf,
  constrained_interactions = NULL,
  positive_interactions = NULL,
  negative_interactions = NULL,
  increasing_interactions = NULL,
  decreasing_interactions = NULL
)
}
\arguments{
\item{X_design}{A numeric design matrix. Each row corresponds to individual
data.}

\item{y}{A numeric observation vector.}

\item{s}{A numeric scalar indicating the maximum order of interaction between
covariates allowed in the estimation method.}

\item{method}{A string indicating the estimation method. One of "em", "hk",
"emhk", "tc", "mars", and "tcmars".}

\item{V}{A numeric scalar. An upper bound on complexity measure (variation).
Required for "hk" and "mars", and possibly used for "emhk" and "tcmars".}

\item{constrained_interactions}{A string vector indicating constrained
interactions between covariates. Possibly used for "emhk" and "tcmars".}

\item{positive_interactions}{A string vector indicating positive interactions
between covariates. Possibly used for "emhk".}

\item{negative_interactions}{A string vector indicating negative interactions
between covariates. Possibly used for "emhk".}

\item{increasing_interactions}{A string vector indicating monotonically
increasing interactions between covariates. Possibly used for "tcmars".}

\item{decreasing_interactions}{A string vector indicating monotonically
decreasing interactions between covariates. Possibly used for "tcmars".}
}
\description{
This function solves the LASSO problem corresponding to an estimation method.
Available estimation methods are entirely monotonic regression ("em"),
Hardy—Krause variation denoising ("hk"), their generalization ("emhk"),
totally convex regression ("tc"), MARS via LASSO ("mars"), and their
generalization ("tcmars"). For details on the corresponding LASSO problems,
see, for example, Section 3 of Fang et al. (2021) (for entirely monotonic
regression and Hardy—Krause variation denoising) and Section 3 and 6 of
Ki et al. (2021) (for MARS via LASSO). There are some other ongoing research
and working papers, and they would be available in the future.
}
\details{
Contrary to entirely monotonic regression (resp. totally convex regression)
where every interaction between covariates is restricted to be positive (resp.
monotonically increasing) and Hardy—Krause variation denoising (resp. MARS via
LASSO) where every interaction is constrained, in their generalization "emhk"
(resp. "tcmars"), you can freely choose which interaction to constrain and
which interaction to restrict to be positive (resp. monotonically increasing)
or negative (resp. monotonically decreasing). When you use any of the
interaction arguments, you need to input a vector of interactions each of
which is represented in a form of concatenation of covariate numbers, sorted
in an increasing order and separated by a hyphen. For example, if you want to
constrain interaction between the second, the third, and the fifth covariates,
you need to include "2-3-5" to your input vector of \code{constrained_interactions}.
}
\examples{
fstar <- function(x) {x[1]**2 + x[2]**2}
X_design <- expand.grid(rep(list(seq(0, 9.0/10, length.out = 10L)), 2L))
theta <- apply(X_design, MARGIN = 1L, FUN = fstar)
sigma <- 1
y <- theta + sigma * rnorm(nrow(X_design))

get_lasso_problem_soln(X_design, y, s = 1L, method = "em")
get_lasso_problem_soln(X_design, y, s = 2L, method = "hk", V = 2)
get_lasso_problem_soln(X_design, y, s = 2L, method = "emhk", V = 1,
                       constrained_interactions = c('1-2'),
                       positive_interactions = c('1'),
                       negative_interactions = c('2'))
get_lasso_problem_soln(X_design, y, s = 2L, method = "tc")
get_lasso_problem_soln(X_design, y, s = 1L, method = "mars", V = 4)
get_lasso_problem_soln(X_design, y, s = 2L, method = "tcmars", V = 2,
                       constrained_interactions = c('1', '1-2'),
                       increasing_interactions = c('2'))
}
\references{
Ki, D., Fang, B., and Guntuboyina, A. (2021). MARS via LASSO.
\url{https://arxiv.org/abs/2111.11694}.

Fang, B., Guntuboyina, A., and Sen, B. (2021). Multivariate
extensions of isotonic regression and total variation denoising via entire
monotonicity and Hardy—Krause variation. \emph{The Annals of Statistics},
\strong{49}(2), 769-792.
}
