% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unique_column_entries.R
\name{get_unique_column_entries}
\alias{get_unique_column_entries}
\title{Get the unique column entries of a design matrix}
\usage{
get_unique_column_entries(X_design, method)
}
\arguments{
\item{X_design}{A numeric design matrix. Each row corresponds to individual
data.}

\item{method}{A string indicating the estimation method and determining the
post-processing. One of "emhk" and "tcmars".}
}
\description{
Given a design matrix, this function returns the list whose ith element is
the sorted vector of the unique entries appearing in the ith column of the
matrix. For entirely monotonic regression, Hardy—Krause variation denoising,
and their generalization ('emhk'), 0 is excluded if exists. For
totally convex regression, MARS via LASSO, and their generalization
('tcmars'), 0 is additionally included and 1 is excluded if exists.
}
