% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_with_hyphen.R
\name{paste_with_hyphen}
\alias{paste_with_hyphen}
\title{Paste two strings with a hyphen}
\usage{
paste_with_hyphen(a, b)
}
\arguments{
\item{a}{A numeric vector. An observation vector.}

\item{b}{A numeric vector. An estimate vector of the same length.}
}
\description{
Given two strings of positive lengths, this function returns the
concatenation of two strings separated by a hyphen. If one of the given two
strings has zero length, this function returns the concatenation without
separation.
}
