% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_regmdc.R
\name{predict_regmdc}
\alias{predict_regmdc}
\title{Give prediction based on a nonparametric regression model with mixed derivative constraints}
\usage{
predict_regmdc(regmdc_model, X_pred)
}
\arguments{
\item{regmdc_model}{An object of the class "regmdc". It is an object returned
by \code{\link{regmdc}}.}

\item{X_pred}{A numeric matrix of new data points. Each row corresponds to
an individual data point at which prediction will be made.}
}
\description{
Given the model built up from \code{\link{regmdc}}, this function provides
prediction at new data points.
}
\examples{
fstar <- function(x) {x[1]**2 + x[2]**2}
X_design <- expand.grid(rep(list(seq(0, 9.0/10, length.out = 10L)), 2L))
theta <- apply(X_design, MARGIN = 1L, FUN = fstar)
sigma <- 1.0
y <- theta + sigma * rnorm(nrow(X_design))

mars_model <- regmdc(X_design, y, s = 1L, method = "mars", V = 4.0)

X_pred <- matrix(c(1.0/3, 2.0/3, 2.0/3, 1.0/3), nrow = 2L)
predict_regmdc(mars_model, X_pred)
}
\seealso{
\code{\link{regmdc}}, which produces nonparametric regression models
with mixed derivative constraints fit to data.
}
