% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_constrained_lasso.R
\name{solve_constrained_lasso}
\alias{solve_constrained_lasso}
\title{Find the solution to the constrained LASSO problem specified below}
\usage{
solve_constrained_lasso(
  y,
  M,
  V = Inf,
  is_sum_constrained_component = NULL,
  is_positive_component = NULL,
  is_negative_component = NULL
)
}
\arguments{
\item{y}{A numeric vector.}

\item{M}{A numeric matrix.}

\item{V}{A numeric scalar.}

\item{is_sum_constrained_component}{A logical vector for whether each
component of x appears in the constraint on the sum of absolute values.}

\item{is_positive_component}{A logical vector for whether each component of x
is restricted to be positive.}

\item{is_negative_component}{A logical vector for whether each component of x
is restricted to be negative.}
}
\description{
Using the conic optimization tools of \code{\link[Rmosek]{mosek}}, this
function solves the problem: min_x \eqn{|| y - M x ||^2} subject to
\itemize{
\item \eqn{\sum_i |x_i| \le V} where the sum is over the components of x where
\code{is_sum_constrained_component} is \code{TRUE},
\item the components of x where \code{is_positive_component} is \code{TRUE} are restricted
to be positive
\item the components of x where \code{is_negative_component} is \code{TRUE} are restricted
to be negative.
}
}
\seealso{
\url{https://docs.mosek.com/9.3/rmosek/tutorial-cqo-shared.html}
for more details about Mosek's conic optimization.
}
