% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_elementary_basis.R
\name{compute_indicator}
\alias{compute_indicator}
\title{Compute the indicator function}
\usage{
compute_indicator(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
Given a numeric vector, this function returns the logical vector indicating
whether each component is nonnegative or not.
}
\examples{
compute_indicator(c(-1.6, 2.3, 0))
}
