% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mse.R
\name{compute_mse}
\alias{compute_mse}
\title{Compute the mean squared error of fitted values}
\usage{
compute_mse(y, y_hat)
}
\arguments{
\item{y}{A numeric observation vector.}

\item{y_hat}{A numeric vector of fitted values.}
}
\description{
Compute the mean squared error of fitted values
}
