% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lasso_matrix.R
\name{get_lasso_matrix}
\alias{get_lasso_matrix}
\title{Construct the matrix for the LASSO problem of an estimation method}
\usage{
get_lasso_matrix(
  X_eval,
  X_design,
  s,
  method,
  is_lattice,
  number_of_bins,
  is_included_basis = NULL
)
}
\arguments{
\item{X_eval}{A numeric evaluation matrix. Each row corresponds to an
individual evaluation point at which basis functions are computed.}

\item{X_design}{A numeric design matrix. Each row corresponds to an
individual data. Basis functions are constructed from this matrix.}

\item{s}{A numeric scalar indicating the maximum order of interaction between
covariates allowed in the estimation method.}

\item{method}{A string indicating the estimation method. One of "em", "hk",
"emhk", "tc", "mars", and "tcmars".}

\item{is_lattice}{A logical scalar for whether the design is lattice or not.
Only used for "em", "hk", and "emhk".}

\item{number_of_bins}{An integer vector of the numbers of bins for the
approximate methods. \code{NULL} if the approximate methods are not used.
Currently available for "tc", "mars", and "tcmars".}

\item{is_included_basis}{A logical vector indicating whether or not each
basis function is included in the LASSO problem.}
}
\description{
Given an estimation method, this function constructs the matrix for the
corresponding LASSO problem. This function also returns the vector indicating
which covariates are used in constructing each basis function. Recall that
basis functions correspond to columns of the matrix for the problem. For
totally concave regression ("tc"), MARS via LASSO ("mars"), and their
generalization ("tcmars"), the indices of columns whose corresponding basis
functions are constrained in the estimation method are additionally returned.
}
\references{
Ki, D., Fang, B., and Guntuboyina, A. (2021). MARS via LASSO.
\url{https://arxiv.org/abs/2111.11694}.

Fang, B., Guntuboyina, A., and Sen, B. (2021). Multivariate
extensions of isotonic regression and total variation denoising via entire
monotonicity and Hardy—Krause variation. \emph{The Annals of Statistics},
\strong{49}(2), 769-792.
}
