% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regmdc.R
\name{regmdc}
\alias{regmdc}
\title{Fit to data a nonparametric regression model with mixed derivative constraints}
\usage{
regmdc(
  X_design,
  y,
  s,
  method,
  V = Inf,
  threshold = 1e-06,
  is_lattice = FALSE,
  is_monotonically_increasing = TRUE,
  is_totally_concave = TRUE,
  number_of_bins = NULL,
  constrained_interactions = NULL,
  positive_interactions = NULL,
  negative_interactions = NULL,
  increasing_interactions = NULL,
  decreasing_interactions = NULL,
  extra_linear_covariates = NULL
)
}
\arguments{
\item{X_design}{A numeric design matrix. Each row corresponds to an
individual data.}

\item{y}{A numeric observation vector of a response variable.}

\item{s}{A numeric scalar indicating the maximum order of interaction between
covariates allowed in the estimation method.}

\item{method}{A string indicating the estimation method. One of "em", "hk",
"emhk", "tc", "mars", and "tcmars".}

\item{V}{A numeric scalar. An upper bound on complexity measure (variation).
Required for "hk" and "mars", and possibly used for "emhk" and "tcmars".}

\item{threshold}{A numeric scalar to determine whether each component of the
solution to the LASSO problem is zero or not.}

\item{is_lattice}{A logical scalar for whether the design is lattice or not.
Only used for "em", "hk", and "emhk".}

\item{is_monotonically_increasing}{A logical scalar for whether the method is
entirely monotonically increasing regression or not (entirely monotonically
decreasing regression). Only used for "em".}

\item{is_totally_concave}{A logical scalar for whether the method is totally
concave regression or not (totally convex regression). Only used for "tc".}

\item{number_of_bins}{An integer or an integer vector of the numbers of bins
for the approximate methods. An integer if the numbers of bins are the same
for all covariates. \code{NULL} if the approximate methods are not used.
Currently available for "tc", "mars", and "tcmars".}

\item{constrained_interactions}{A string vector indicating constrained
interactions between covariates. Possibly used for "emhk" and "tcmars". See
details below.}

\item{positive_interactions}{A string vector indicating positive interactions
between covariates. Possibly used for "emhk". See details below.}

\item{negative_interactions}{A string vector indicating negative interactions
between covariates. Possibly used for "emhk". See details below.}

\item{increasing_interactions}{A string vector indicating monotonically
increasing interactions between covariates. Possibly used for "tcmars". See
details below.}

\item{decreasing_interactions}{A string vector indicating monotonically
decreasing interactions between covariates. Possibly used for "tcmars". See
details below.}

\item{extra_linear_covariates}{An integer vector or a string vector of extra
linear covariates added to the model. Possibly used for "tc", "mars", and
"tcmars". See details below.}
}
\description{
Given an estimation method, this function builds up a model by solving the
corresponding constrained LASSO problem. Available estimation methods are
entirely monotonic regression ("em"), Hardy—Krause variation denoising ("hk"),
their generalization ("emhk"), totally concave regression ("tc"), MARS via
LASSO ("mars"), and their generalization ("tcmars"). For details on the
corresponding LASSO problems, see, for example, Section 3 of Fang et al.
(2021) (for entirely monotonic regression and Hardy—Krause variation
denoising) and Section 2 of Ki et al. (2024+) (for MARS via LASSO). There are
some other ongoing research and working papers, and they will be available in
the future.
}
\details{
Contrary to entirely monotonic regression (resp. totally concave regression)
where every interaction between covariates is restricted to be positive (resp.
monotonically increasing) and Hardy—Krause variation denoising (resp. MARS via
LASSO) where every interaction is constrained, in their generalization "emhk"
(resp. "tcmars"), you can freely choose which interaction to constrain and
which interaction to restrict to be positive (resp. monotonically increasing)
or negative (resp. monotonically decreasing). When you use any of the
interaction arguments, you need to input a vector of interactions each of
which is represented in a form of concatenation of covariate numbers, sorted
in an increasing order and separated by a hyphen. For example, if you want to
constrain interaction between the second, the third, and the fifth covariates,
you need to include "2-3-5" to your input vector of \code{constrained_interactions}.

If \code{number_of_bins} is not \code{NULL}, then the approximate methods are used.
Refer to \code{\link{get_unique_column_entries}} to see what differences are
made in the approximate methods. The approximate methods for totally concave
regression, MARS via LASSO, and their generalization have been implemented.
Approximate methods for entirely monotonic regression, Hardy—Krause variation
denoising, and their generalization will be available in the future.

You can also add to the model extra covariates that have linear relationship
with the response variable. The interaction between these covariates and the
interaction between these covariates and the others are not considered in the
model. You can put an integer vector of column indices or a string vector of
column names into \code{extra_linear_covariates}.
}
\examples{
fstar <- function(x) {x[1]**2 + x[2]**2}
X_design <- expand.grid(rep(list(seq(0, 13.0/14, length.out = 14L)), 2L))
theta <- apply(X_design, MARGIN = 1L, FUN = fstar)
sigma <- 1.0
y <- theta + sigma * rnorm(nrow(X_design))

regmdc(X_design, y, s = 1L, method = "em")
regmdc(X_design, y, s = 2L, method = "em")
regmdc(X_design, y, s = 2L, method = "em", is_lattice = TRUE)
regmdc(X_design, y, s = 2L, method = "em",
       is_monotonically_increasing = FALSE)
regmdc(X_design, y, s = 2L, method = "hk", V = 2.0)
regmdc(X_design, y, s = 2L, method = "emhk", V = 1.0,
       constrained_interactions = c('1-2'),
       positive_interactions = c('1'),
       negative_interactions = c('2'))
regmdc(X_design, y, s = 2L, method = "tc")
regmdc(X_design, y, s = 2L, method = "tc", is_totally_concave = FALSE)
regmdc(X_design, y, s = 2L, method = "mars", V = 4.0)
regmdc(X_design, y, s = 2L, method = "tcmars", V = 2.0,
       constrained_interactions = c('1', '1-2'),
       increasing_interactions = c('2'))

X_design <- cbind(runif(20), runif(20), runif(20))
colnames(X_design) <- c("VarA", "VarB", "VarC")
theta <- apply(X_design, MARGIN = 1L, FUN = fstar)
sigma <- 1.0
y <- theta + sigma * rnorm(nrow(X_design))
regmdc(X_design, y, s = 2L, method = "tc", extra_linear_covariates = c(3L))
regmdc(X_design, y, s = 2L, method = "tc", extra_linear_covariates = c(2L, 3L))
regmdc(X_design, y, s = 2L, method = "tc", extra_linear_covariates = c("VarC"))
regmdc(X_design, y, s = 2L, method = "mars", V = 4.0, number_of_bins = 10L)
regmdc(X_design, y, s = 2L, method = "mars", V = 4.0, number_of_bins = c(5L, 10L, 10L))
regmdc(X_design, y, s = 2L, method = "mars", V = 4.0, number_of_bins = c(5L, 10L, NA))
regmdc(X_design, y, s = 2L, method = "mars", V = 4.0, number_of_bins = c(5L, 10L, NA),
       extra_linear_covariates = c("VarC"))
}
\references{
Ki, D., Fang, B., and Guntuboyina, A. (2024+). MARS via LASSO.
Accepted at \emph{Annals of Statistics}. Available at
\url{https://arxiv.org/abs/2111.11694}.

Fang, B., Guntuboyina, A., and Sen, B. (2021). Multivariate
extensions of isotonic regression and total variation denoising via entire
monotonicity and Hardy—Krause variation. \emph{Annals of Statistics},
\strong{49}(2), 769-792.
}
