% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_constrained_lasso.R
\name{solve_constrained_lasso}
\alias{solve_constrained_lasso}
\title{Find the solution to the constrained LASSO problem specified below}
\usage{
solve_constrained_lasso(
  y,
  M,
  V = Inf,
  sum_constrained_indices = NULL,
  positive_indices = NULL,
  negative_indices = NULL
)
}
\arguments{
\item{y}{A numeric vector.}

\item{M}{A numeric matrix.}

\item{V}{A numeric scalar.}

\item{sum_constrained_indices}{A numeric vector. The indices of the
components of x whose sum of absolute values are constrained.}

\item{positive_indices}{A numeric vector. The indices of the components of x
that are restricted to be positive.}

\item{negative_indices}{A numeric vector. The indices of the components of x
that are restricted to be negative.}
}
\description{
Using the conic optimization tools of \code{\link[Rmosek]{mosek}}, this
function solves the problem: min_x \eqn{|| y - M x ||^2} subject to
\itemize{
\item \eqn{\sum_i |x_i| \le V} where the sum is over the components of x indexed
by \code{sum_constrained_indices},
\item the components of x indexed by \code{positive_indices} are restricted to be
positive,
\item the components of x indexed by \code{negative_indices} are restricted to be
negative.
}
}
\seealso{
\url{https://docs.mosek.com/9.3/rmosek/tutorial-cqo-shared.html}
for more details about Mosek's conic optimization.
}
